Avoid spamming the player with seek commands
The player will sometimes get stuck buffering for many seconds. I think this
is because we seek too often; while the player is trying to load data for
one timestamp, we're seeking to the next one. This change is kind of
experimental, because the bug only repros intermittently, but I am hopeful
that this will fix it.